Change meta in Item to short

+++ src/main/java/cn/nukkit/block/Block.java
@@ -239,7 +239,7 @@ public class Block extends Position implements Metadatable, Cloneable {
     public static boolean[] transparent = null;
 
     protected int id;
-    protected int meta = 0;
+    protected short meta = 0;
 
     public AxisAlignedBB boundingBox = null;
 
@@ -249,7 +249,7 @@ public class Block extends Position implements Metadatable, Cloneable {
 
     public Block(int id, Integer meta) {
         this.id = id;
-        this.meta = meta != null ? meta : 0;
+        this.meta = (short) (meta != null ? meta : 0);
     }
 
     public static void init() {
@@ -433,12 +433,12 @@ public class Block extends Position implements Metadatable, Cloneable {
 
     }
 
-    public final int getDamage() {
+    public final short getDamage() {
         return this.meta;
     }
 
     public final void setDamage(Integer meta) {
-        this.meta = meta == null ? 0 : meta & 0x0f;
+        this.meta = (short) (meta == null ? 0 : meta & 0x0f);
     }
 
     final public void position(Position v) {
